This is the solution of time conversion question from hackerrank.
Given a time in 12-hour AM/PM format, convert it to military (24-hour) time.

Note: Midnight is 12:00:00AM on a 12-hour clock, and 00:00:00 on a 24-hour clock. Noon is 12:00:00PM on a 12-hour clock, and 12:00:00 on a 24-hour clock.

Function Description

Complete the timeConversion function in the editor below. It should return a new string representing the input time in 24 hour format.

timeConversion has the following parameter(s):

s: a string representing time in 12 hour format
Output Format

Convert and print the given time in 24-hour format, where 00<=hh<=23 .

Sample Input 0

07:05:45PM
Sample Output 0

19:05:45